reorganized optimize parameters
